+++ stream-chat-android-compose-sample/src/main/java/io/getstream/chat/android/compose/sample/ui/MessagesActivity.kt
@@ -24,6 +24,7 @@ import androidx.activity.viewModels
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.interaction.MutableInteractionSource
+import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
@@ -61,6 +62,9 @@ import io.getstream.chat.android.compose.ui.messages.attachments.AttachmentsPick
 import io.getstream.chat.android.compose.ui.messages.composer.MessageComposer
 import io.getstream.chat.android.compose.ui.messages.list.MessageList
 import io.getstream.chat.android.compose.ui.theme.ChatTheme
+import io.getstream.chat.android.compose.ui.theme.MessageComposerTheme
+import io.getstream.chat.android.compose.ui.theme.StreamColors
+import io.getstream.chat.android.compose.ui.theme.StreamTypography
 import io.getstream.chat.android.compose.ui.util.rememberMessageListState
 import io.getstream.chat.android.compose.viewmodel.messages.AttachmentsPickerViewModel
 import io.getstream.chat.android.compose.viewmodel.messages.MessageComposerViewModel
@@ -86,6 +90,7 @@ class MessagesActivity : BaseConnectedActivity() {
             context = this,
             channelId = requireNotNull(intent.getStringExtra(KEY_CHANNEL_ID)),
             autoTranslationEnabled = ChatApp.autoTranslationEnabled,
+            isComposerLinkPreviewEnabled = ChatApp.isComposerLinkPreviewEnabled,
             deletedMessageVisibility = DeletedMessageVisibility.ALWAYS_VISIBLE,
             messageId = intent.getStringExtra(KEY_MESSAGE_ID),
             parentMessageId = intent.getStringExtra(KEY_PARENT_MESSAGE_ID),
@@ -101,10 +106,24 @@ class MessagesActivity : BaseConnectedActivity() {
         super.onCreate(savedInstanceState)
 
         setContent {
+            val colors = if (isSystemInDarkTheme()) StreamColors.defaultDarkColors() else StreamColors.defaultColors()
+            val typography = StreamTypography.defaultTypography()
             ChatTheme(
+                colors = colors,
+                typography = typography,
                 dateFormatter = ChatApp.dateFormatter,
                 autoTranslationEnabled = ChatApp.autoTranslationEnabled,
+                isComposerLinkPreviewEnabled = ChatApp.isComposerLinkPreviewEnabled,
                 allowUIAutomationTest = true,
+                messageComposerTheme = MessageComposerTheme.defaultTheme(typography).let { messageComposerTheme ->
+                    messageComposerTheme.copy(
+                        attachmentCancelIcon = messageComposerTheme.attachmentCancelIcon.copy(
+                            painter = painterResource(id = R.drawable.stream_compose_ic_clear),
+                            tint = colors.overlayDark,
+                            backgroundColor = colors.appBackground,
+                        ),
+                    )
+                },
             ) {
                 MessagesScreen(
                     viewModelFactory = factory,
@@ -330,6 +349,7 @@ class MessagesActivity : BaseConnectedActivity() {
     }
 
     companion object {
+        private const val TAG = "MessagesActivity"
         private const val KEY_CHANNEL_ID = "channelId"
         private const val KEY_MESSAGE_ID = "messageId"
         private const val KEY_PARENT_MESSAGE_ID = "parentMessageId"