+++ stream-chat-android-docs/src/main/kotlin/io/getstream/chat/docs/kotlin/cookbook/CookbookMainActivity.kt
@@ -1,0 +1,91 @@
+package io.getstream.chat.docs.kotlin.cookbook
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.lifecycle.lifecycleScope
+import androidx.navigation.compose.NavHost
+import androidx.navigation.compose.composable
+import androidx.navigation.compose.rememberNavController
+import io.getstream.chat.android.compose.ui.theme.ChatTheme
+import io.getstream.chat.docs.kotlin.cookbook.ui.CustomChannelListScreen
+import io.getstream.chat.docs.kotlin.cookbook.ui.CustomComposerAndAttachmentsPicker
+import io.getstream.chat.docs.kotlin.cookbook.ui.CustomMessageListHeader
+import io.getstream.chat.docs.kotlin.cookbook.ui.CustomMessageListScreen
+import io.getstream.chat.docs.kotlin.cookbook.ui.theme.CookbookTheme
+import io.getstream.chat.docs.kotlin.cookbook.utils.connectUser
+import io.getstream.chat.docs.kotlin.cookbook.utils.initChatClient
+import io.getstream.chat.docs.kotlin.cookbook.utils.userCredentials
+import kotlinx.coroutines.launch
+
+class CookbookMainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        init()
+
+        setContent {
+            val navController = rememberNavController()
+
+            CookbookTheme {
+                NavHost(
+                    navController = navController,
+                    startDestination = AppScreens.CustomChannelList.route
+                ) {
+                    composable(AppScreens.CustomChannelList.route) {
+                        CustomChannelListScreen(
+                            navigateToMessageList = { cid ->
+                                // navController.navigate(AppScreens.CustomMessageList.routeWithArg(cid))
+                                // navController.navigate(AppScreens.CustomMessageListHeader.routeWithArg(cid))
+                                navController.navigate(AppScreens.CustomMessageComposer.routeWithArg(cid))
+                            }
+                        )
+                    }
+
+                    composable(AppScreens.CustomMessageList.route) { backStackEntry ->
+                        CustomMessageListScreen(
+                            cid = backStackEntry.arguments?.getString("cid")
+                        )
+                    }
+
+                    composable(AppScreens.CustomMessageListHeader.route) { backStackEntry ->
+                        ChatTheme {
+                            CustomMessageListHeader(
+                                cid = backStackEntry.arguments?.getString("cid"),
+                                onBackClick = { navController.popBackStack() }
+                            )
+                        }
+                    }
+
+                    composable(AppScreens.CustomMessageComposer.route) { backStackEntry ->
+                        ChatTheme {
+                            CustomComposerAndAttachmentsPicker(
+                                cid = backStackEntry.arguments?.getString("cid"),
+                                onBackClick = { navController.popBackStack() }
+                            )
+                        }
+                    }
+                }
+            }
+        }
+    }
+
+    private fun init() {
+        initChatClient(userCredentials.apiKey, this.applicationContext)
+        lifecycleScope.launch { connectUser() }
+    }
+}
+
+enum class AppScreens(val route: String) {
+    CustomChannelList("channel_list"),
+    CustomMessageList("message_list/{cid}"),
+    CustomMessageListHeader("message_header/{cid}"),
+    CustomMessageComposer("message_composer/{cid}");
+
+    fun routeWithArg(argValue: Any): String = when (this) {
+        CustomMessageList -> this.route.replace("{cid}", argValue.toString())
+        CustomMessageListHeader -> this.route.replace("{cid}", argValue.toString())
+        CustomMessageComposer -> this.route.replace("{cid}", argValue.toString())
+        else -> this.route
+    }
+}