+++ CHANGELOG.md
@@ -8,8 +8,19 @@
 ### ✅ Added
 
 ### ⚠️ Changed
-
 - Bump Compose UI to 1.6.2 and Compose compiler to 1.5.10.
+
+### ❌ Removed
+
+## stream-chat-android-core
+### 🐞 Fixed
+
+### ⬆️ Improved
+
+### ✅ Added
+- Added `LinkPreview` model to represent the link preview data. [#5184](https://github.com/GetStream/stream-chat-android/pull/5184)
+
+### ⚠️ Changed
 
 ### ❌ Removed
 
@@ -17,11 +28,14 @@
 ### 🐞 Fixed
 - Prevent inserting reaction which violates `ForeignKey` constraint in local DB. [#5164](https://github.com/GetStream/stream-chat-android/pull/5164)
 - Fixed crash in `ChatEventsObservable.onNext`. [#5165](https://github.com/GetStream/stream-chat-android/pull/5165)
+- Fixed `ArrayIndexOutOfBoundsException` in `ApiRequestsDumper`. [#5187](https://github.com/GetStream/stream-chat-android/pull/5187)
 
 ### ⬆️ Improved
+- Ensure fresh token is used to establish WS connection. [#5185](https://github.com/GetStream/stream-chat-android/pull/5185)
 
 ### ✅ Added
 - Added `DeleteChannelListener`. [#5164](https://github.com/GetStream/stream-chat-android/pull/5164)
+- Added `ChatClient.enrichUrl` to enrich the URL with the preview data. [#5184](https://github.com/GetStream/stream-chat-android/pull/5184)
 
 ### ⚠️ Changed
 
@@ -63,7 +77,8 @@
 
 ### ⚠️ Changed
 - 🚨Breaking change: Changed `MessagePositionHandler.handleMessagePosition` signature. [#5168](https://github.com/GetStream/stream-chat-android/pull/5168)
-  * Added `isInThread: Boolean` parameter. 
+  * Added `isInThread: Boolean` parameter.
+- Deprecated `AttachmentSelectionListener` class, use `AttachmentsSelectionListener` instead. [#5178](https://github.com/GetStream/stream-chat-android/pull/5178)
 
 ### ❌ Removed
 
@@ -89,6 +104,7 @@
   * `OnReactionViewClickListener` and `MessageListView.setOnReactionViewClickListener`
   * `OnUserReactionClickListener` and `MessageListView.setOnUserReactionClickListener`
 - Added `messageBuilder` parameter to `MessageComposerViewModel.bindView` to allow customizing the message builder. [#5169](https://github.com/GetStream/stream-chat-android/pull/5169)
+- Added `MessageComposerView.attachmentsPickerDialogBuilder` lambda that allow you to create your own Picker Dialog. [#5178](https://github.com/GetStream/stream-chat-android/pull/5178)
 
 ### ⚠️ Changed
 
@@ -101,6 +117,8 @@
 ### ⬆️ Improved
 
 ### ✅ Added
+- Added `MessageComposerTheme` to customize the message composer. [#5183](https://github.com/GetStream/stream-chat-android/pull/5183)
+- Added `ComposerLinkPreview` to show link previews in the message composer. [#5184](https://github.com/GetStream/stream-chat-android/pull/5184)
 
 ### ⚠️ Changed